+++ taxtastic/subcommands/update_taxids.py
@@ -19,7 +19,7 @@ database. Use in preparation for ``taxit taxtable``. Takes sequence
 info file as passed to ``taxit create --seq-info``
 
 """
-
+import codecs
 import csv
 import logging
 import sys
@@ -33,33 +33,46 @@ log = logging.getLogger(__name__)
 
 
 def build_parser(parser):
-    parser.add_argument(
-        'infile', type=taxtastic.utils.Opener('r'),
-        help=('Input CSV file to process, minimally containing the field '
-              '`tax_id`. Use "-" for stdin.'))
     parser = taxtastic.utils.add_database_args(parser)
     parser.add_argument(
-        '-o', '--outfile',
-        default=sys.stdout,
-        type=taxtastic.utils.Opener('wt'),
-        help='Modified version of input file [default: stdout]')
-    input_format = parser.add_mutually_exclusive_group(required=False)
-    input_format.add_argument(
-        '--taxid-column', default='tax_id',
-        help='name of column containing tax_ids to be replaced [%(default)s]')
-    input_format.add_argument(
-        '--tax-id-file', action='store_true',
+        'infile',
+        type=taxtastic.utils.Opener('rt'),
+        help='Input file with taxids. Use "-" for stdin.')
+    parser.add_argument(
+        '--delimiter',
+        default=',',
+        metavar='',
+        type=lambda x: codecs.decode(str(x), 'unicode_escape'),
+        help='Infile columns delimiter [%(default)s]')
+    parser.add_argument(
+        '--taxid-column',
+        default='tax_id',
+        metavar='',
+        help='name of column or index if headerless containing '
+             'tax_ids to be replaced [%(default)s]')
+    parser.add_argument(
+        '--tax-id-file',
+        action='store_true',
         help='Infile is a headerless text file '
              'of tax_ids separated by newlines. [%(default)s]')
     parser.add_argument(
-        '--unknowns', type=taxtastic.utils.Opener('wt'),
-        help=('optional output file containing rows with unknown tax_ids '
-              'having no replacements in merged table'))
+        '--unknowns',
+        metavar='',
+        type=taxtastic.utils.Opener('wt'),
+        help='optional output file containing rows with unknown tax_ids '
+             'having no replacements in merged table')
     parser.add_argument(
-        '-a', '--unknown-action', choices=['drop', 'ignore', 'error'],
+        '-a', '--unknown-action',
+        choices=['drop', 'ignore', 'error'],
         default='error',
-        help=('action to perform for tax_ids with no replacement '
-              'in merged table [%(default)s]'))
+        help='action to perform for tax_ids with no replacement '
+             'in merged table [%(default)s]')
+    parser.add_argument(
+        '-o', '--outfile',
+        default=sys.stdout,
+        metavar='',
+        type=taxtastic.utils.Opener('wt'),
+        help='Modified version of input file [stdout]')
 
     # not implemented for now
     # parser.add_argument(
@@ -76,6 +89,7 @@ def action(args):
         reader = csv.DictReader(args.infile, fieldnames=[args.taxid_column])
     else:
         reader = csv.DictReader(args.infile)
+
     fieldnames = reader.fieldnames
     taxid_column = args.taxid_column
     drop = args.unknown_action == 'drop'
@@ -83,14 +97,21 @@ def action(args):
     ignore = args.unknown_action == 'ignore'
 
     if taxid_column not in fieldnames:
-        raise ValueError("No column " + args.taxid_column)
+        if taxid_column.isnumeric():
+            index = int(taxid_column)
+            taxid_column = list(fieldnames)[index - 1]
+        else:
+            raise ValueError("No column " + args.taxid_column)
 
     # TODO: remove unless --use-names is implemented
     # if args.use_names and args.name_column not in fieldnames:
     #     raise ValueError("No column " + args.name_column)
 
-    writer = csv.DictWriter(args.outfile, fieldnames=fieldnames)
-    if not args.tax_id_file:
+    writer = csv.DictWriter(
+        args.outfile,
+        delimiter=args.delimiter,
+        fieldnames=fieldnames)
+    if not (args.tax_id_file or args.taxid_column.isnumeric()):
         writer.writeheader()
 
     if args.unknowns:
@@ -104,8 +125,8 @@ def action(args):
 
     with tax.engine.connect() as con:
         log.info('reading table merged')
-        result = con.execute(sa.text(
-            'select old_tax_id, new_tax_id from {merged}'.format(**tax.tables)))
+        q = 'select old_tax_id, new_tax_id from {merged}'.format(**tax.tables)
+        result = con.execute(sa.text(q))
         mergedict = dict(result.fetchall())
 
         log.info('reading tax_ids from table nodes')