+++ CHANGES.rst
@@ -7,6 +7,7 @@ Upcoming
 
 * Updated NCBI taxdmp.zip download protocol from ftp to https [GH-156]
 * New subcommand `taxit named` that filters unclassified, unnamed taxonomy ids
+* Subcommand update_taxids can now process a tab delimited headerless file
 
 0.10.1
 ======